Use correct format for size_t
The format for some size_t variables were recently changed from %d to
%llu. This fixed a problem on 64 bit architectures, since there size_t
is 64 bits wide and %d is the format for a 32 bit type and %llu is the
format for a 64 bit type. However, this change introduced a regression
on 32 bit architectures where size_t is 32 bits wide.

This change caused some of the tests to fail with a segmentation fault
on the Debian armhf architecture:

The following tests FAILED:
	108 - XRootD::noauth::test (Failed)
	111 - XRootD::host::test (Failed)
	114 - XRootD::unix::test (Failed)
	117 - XRootD::sss::test (Failed)
	120 - XRootD::http::test (Failed)

This commit adds a type cast so that the new format works for all
architectures.

+++ src/XrdCl/XrdClCopyProcess.cc
@@ -369,7 +369,7 @@ namespace XrdCl
     std::vector<PropertyList>::iterator it;
 
     log->Debug( UtilityMsg, "CopyProcess: %llu jobs to prepare",
-                pImpl->pJobProperties.size() );
+                (unsigned long long) pImpl->pJobProperties.size() );
 
     std::map<std::string, uint32_t> targetFlags;
     int i = 0;
