+++ src/XrdCl/XrdClJobManager.cc
@@ -83,7 +83,7 @@ namespace XrdCl
       }
     }
     pRunning = true;
-    log->Debug( JobMgrMsg, "Job manager started, %llu workers", pWorkers.size() );
+    log->Debug( JobMgrMsg, "Job manager started, %llu workers", (unsigned long long) pWorkers.size() );
     return true;
   }
